test/smoke/requirements.txt
bitcoincash==0.1.5
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.3
deepdiff==4.2.0
idna==2.8
ordered-set==3.1.1
python-bitcointx==1.1.1.post2
python-litecointx==0.1.2
python-dogecointx==0.1.0
web3==5.9.0
rlp==1.2.0
requests==2.22.0
six==1.14.0
tenacity==6.1.0
urllib3==1.25.8
tqdm==4.45.0
websocket-client==0.57.0
bech32==1.2.0
terra-sdk==2.0.4
terra-proto==1.0.1
durationpy==0.5
